Add Post method for Car

+++ CarRentalManagerAPI/Entities/Car.cs
@@ -12,7 +12,7 @@ namespace CarRentalManagerAPI.Entities
         public int EnginePower { get; set; }
         public DrivingLicenseCategoryEnum DrivingLicenseCategory { get; set; }
         public int NumberOfSeats { get; set; }
-        public int PricePerDay { get; set; }
+        public double PricePerDay { get; set; }
         public string RegistrationNumber { get; set; }
         public string VIN { get; set; }
         public CarStatusEnum Status { get; set; }
